+++ classifai-core/src/main/java/ai/classifai/database/annotation/bndbox/BoundingBoxDbQuery.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package ai.classifai.database.annotation.boundingboxdb;
+package ai.classifai.database.annotation.bndbox;
 
 import ai.classifai.database.annotation.AnnotationQuery;
