+++ classifai-core/src/main/java/ai/classifai/selector/annotation/ToolFolderSelector.java
@@ -16,8 +16,8 @@
 package ai.classifai.selector.annotation;
 
 import ai.classifai.data.type.image.ImageFileType;
-import ai.classifai.database.loader.ProjectLoader;
-import ai.classifai.ui.button.LookFeelSetter;
+import ai.classifai.loader.ProjectLoader;
+import ai.classifai.ui.component.LookFeelSetter;
 import ai.classifai.ui.launcher.LogoLauncher;
 import ai.classifai.selector.filesystem.FileSystemStatus;
 import ai.classifai.util.ParamConfig;
@@ -40,7 +40,7 @@ import java.util.ArrayList;
  * @author codenamewei
  */
 @Slf4j
-public class FolderSelector{
+public class ToolFolderSelector{
 
     private static FileNameExtensionFilter imgfilter = new FileNameExtensionFilter("Image Files", ImageFileType.getImageFileTypes());
