+++ classifai-core/src/main/java/ai/classifai/ui/component/LookFeelSetter.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package ai.classifai.ui.button;
+package ai.classifai.ui.component;
 
 import com.formdev.flatlaf.FlatDarkLaf;
 import lombok.extern.slf4j.Slf4j;
